third_party/WebKit/LayoutTests/platform/mac/external/wpt/html/semantics/tabular-data/processing-model-1/span-limits-expected.txt

This is a testharness.js-based test.
PASS colspan of 1000 must work 
FAIL colspan of 1001 must be treated as 1000 assert_equals: expected 23 but got 37
PASS rowspan of 65534 must work 
PASS rowspan of 65535 must be treated as 65534 
Harness: the test ran to completion.
